requirements.txt
streamlit>=1.36.0
google-genai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tinycss2>=1.2.0
numpy>=1.24.0
